Crossterm.jl: A Cross-platform Terminal Manipulation Library in Julia

Crossterm is a Julia terminal manipulation library designed to write cross-platform text-based interfaces based on crossterm-rs/crossterm. It supports all UNIX and Windows terminals down to Windows 7.

The Julia interface provides simpler interface with functions like:

hide()

show()

up()

down()

left()

right()

next()

prev()

Crossterm.alternate_screen

Crossterm.attribute

Crossterm.blinking

Crossterm.bracketed_paste

Crossterm.clear

Crossterm.clear_last_error

Crossterm.column

Crossterm.down

Crossterm.focus_change

Crossterm.hide

Crossterm.keyboard_enchancement_flags

Crossterm.last_error_message

Crossterm.left

Crossterm.line_wrap

Crossterm.mouse_capture

Crossterm.next

Crossterm.poll

Crossterm.position

Crossterm.position

Crossterm.prev

Crossterm.raw_mode

Crossterm.read

Crossterm.reset_color

Crossterm.restore

Crossterm.right

Crossterm.row

Crossterm.save

Crossterm.scroll_down

Crossterm.scroll_up

Crossterm.show

Crossterm.size

Crossterm.size

Crossterm.style

Crossterm.title

Crossterm.to

Crossterm.up

Crossterm.alternate_screen — Function

Enter or leave the alternate screen buffer of the terminal.

Crossterm.attribute — Method

Set the text attribute in the terminal.

Crossterm.blinking — Function

Enable or disable cursor blinking in the terminal.

Crossterm.bracketed_paste — Function

Enable or disable bracketed paste mode in the terminal.

Crossterm.clear — Function

Clear the terminal screen or a part of it.

Crossterm.clear_last_error — Method

Clear the last error message stored by the crossterm library.

Crossterm.column — Function

Move the terminal cursor to the column.

Crossterm.down — Function

Move the terminal cursor down by rows number of rows.

Crossterm.focus_change — Function

Enable or disable focus change event reporting in the terminal.

Crossterm.hide — Method

Hide the terminal cursor.

Crossterm.keyboard_enchancement_flags — Function

Push or pop the keyboard enhancement flags.

Crossterm.last_error_message — Method

Return the most recent error message from the crossterm library.

Crossterm.left — Function

Move the terminal cursor left by columns number of columns.

Crossterm.line_wrap — Function

Enable or disable line wrapping in the terminal.

Crossterm.mouse_capture — Function

Enable or disable mouse capture event reporting in the terminal.

Crossterm.next — Function

Move the terminal cursor to the next line n times.

Crossterm.poll — Function

Poll the terminal for input events.

Crossterm.position — Method

Set the position of the terminal cursor to p coordinates.

Crossterm.position — Method

Get the current position of the terminal cursor.

Crossterm.prev — Function

Move the terminal cursor to the previous line n times.

Crossterm.raw_mode — Function

Enable or disable raw mode in the terminal.

Crossterm.read — Method

Read an event from the terminal.

Crossterm.reset_color — Method

Reset terminal colors to default.

Crossterm.restore — Method

Restore the terminal cursor to the last saved position.

Crossterm.right — Function

Move the terminal cursor right by columns number of columns.

Crossterm.row — Function

Move the terminal cursor to the row.

Crossterm.save — Method

Save the current position of the terminal cursor.

Crossterm.scroll_down — Function

Scroll the terminal screen down by n lines.

Crossterm.scroll_up — Function

Scroll the terminal screen up by n lines.

Crossterm.show — Method

Show the terminal cursor.

Crossterm.size — Method

Set the size of the terminal.

Crossterm.size — Method

Get the size of the terminal.

Crossterm.style — Method

Set the style of the terminal cursor.

Crossterm.title — Method

Set the title of the terminal.

Crossterm.to — Method

Move the terminal cursor to the specified (x, y) coordinates.

Crossterm.up — Function

Move the terminal cursor up by rows number of rows.