Crossterm.jl: A Cross-platform Terminal Manipulation Library in Julia

Crossterm is a Julia terminal manipulation library designed to write cross-platform text-based interfaces based on crossterm-rs/crossterm. It supports all UNIX and Windows terminals down to Windows 7.

The Julia interface provides simpler interface with functions like:

hide()

show()

up()

down()

left()

right()

next()

prev()

and more.

See API documentation below.

Crossterm.alternate_screen — Function

Enter or leave the alternate screen buffer of the terminal.

Crossterm.attribute — Method

Set the text attribute in the terminal.

Crossterm.blinking — Function

Enable or disable cursor blinking in the terminal.

Crossterm.bracketed_paste — Function

Enable or disable bracketed paste mode in the terminal.

Crossterm.clear — Function

Clear the terminal screen or a part of it.

Crossterm.clear_last_error — Method

Clear the last error message stored by the crossterm library.

Crossterm.column — Function

Move the terminal cursor to the column.

Crossterm.cursor — Function

Show or hide cursor in the terminal.

Crossterm.down — Function

Move the terminal cursor down by rows number of rows.

Crossterm.focus_change — Function

Enable or disable focus change event reporting in the terminal.

Crossterm.hide — Method

Hide the terminal cursor.

Crossterm.keyboard_enhancement_flags — Function

Push or pop the keyboard enhancement flags.

Crossterm.last_error_message — Method

Return the most recent error message from the crossterm library.

Crossterm.left — Function

Move the terminal cursor left by columns number of columns.

Crossterm.line_wrap — Function

Enable or disable line wrapping in the terminal.

Crossterm.mouse_capture — Function

Enable or disable mouse capture event reporting in the terminal.

Crossterm.next — Function

Move the terminal cursor to the next line n times.

Crossterm.poll — Function

Poll the terminal for input events for duration (Dates.Period)

Crossterm.poll — Method

Poll the terminal for input events for duration (seconds)

Crossterm.position — Method

Set the position of the terminal cursor to p coordinates.

Crossterm.position — Method

Get the current position of the terminal cursor.

Crossterm.prev — Function

Move the terminal cursor to the previous line n times.

Crossterm.raw_mode — Function

Enable or disable raw mode in the terminal.

Crossterm.read — Method

Read an event from the terminal.

Crossterm.reset_color — Method

Reset terminal colors to default.

Crossterm.restore — Method

Restore the terminal cursor to the last saved position.

Crossterm.right — Function

Move the terminal cursor right by columns number of columns.

Crossterm.row — Function

Move the terminal cursor to the row.

Crossterm.save — Method

Save the current position of the terminal cursor.

Crossterm.scroll_down — Function

Scroll the terminal screen down by n lines.

Crossterm.scroll_up — Function

Scroll the terminal screen up by n lines.

Crossterm.show — Method

Show the terminal cursor.

Crossterm.size — Method

Set the size of the terminal.

Crossterm.size — Method

Get the size of the terminal.

Crossterm.style — Method

Set the style of the terminal cursor.

Crossterm.title — Method

Set the title of the terminal.

Crossterm.to — Method

Move the terminal cursor to the specified (x, y) coordinates.

Crossterm.up — Function

Move the terminal cursor up by rows number of rows.

Full C-API documentation is below

Crossterm.LibCrossterm.crossterm_ClearType — Type

crossterm_ClearType

Different ways to clear the terminal buffer.

Crossterm.LibCrossterm.crossterm_CursorPosition — Type

crossterm_CursorPosition

CursorPosition struct

Crossterm.LibCrossterm.crossterm_CursorStyle — Type

crossterm_CursorStyle

Style of the cursor. It uses two types of escape codes, one to control blinking, and the other the shape.

Crossterm.LibCrossterm.crossterm_Event_Tag — Type

crossterm_Event_Tag

Represents an event.

Crossterm.LibCrossterm.crossterm_KeyCode_Tag — Type

crossterm_KeyCode_Tag

Represents a key.

Crossterm.LibCrossterm.crossterm_KeyEvent — Type

crossterm_KeyEvent

Represents a key event.

Crossterm.LibCrossterm.crossterm_KeyEventKind — Type

crossterm_KeyEventKind

Represents a keyboard event kind.

Crossterm.LibCrossterm.crossterm_KeyEventState — Type

crossterm_KeyEventState

Represents extra state about the key event.

Note: This state can only be read if [KeyboardEnhancementFlags::DISAMBIGUATE\_ESCAPE\_CODES] has been enabled with [PushKeyboardEnhancementFlags].

Crossterm.LibCrossterm.crossterm_KeyModifiers — Type

crossterm_KeyModifiers

Represents key modifiers (shift, control, alt, etc.).

Note: SUPER, HYPER, and META can only be read if [KeyboardEnhancementFlags::DISAMBIGUATE\_ESCAPE\_CODES] has been enabled with [PushKeyboardEnhancementFlags].

Crossterm.LibCrossterm.crossterm_MediaKeyCode — Type

crossterm_MediaKeyCode

Represents a media key (as part of [KeyCode::Media]).

Crossterm.LibCrossterm.crossterm_ModifierKeyCode — Type

crossterm_ModifierKeyCode

Represents a modifier key (as part of [KeyCode::Modifier]).

Crossterm.LibCrossterm.crossterm_MouseButton — Type

crossterm_MouseButton

Represents a mouse button.

Crossterm.LibCrossterm.crossterm_MouseEvent — Type

crossterm_MouseEvent

Represents a mouse event.

Platform-specific Notes

Mouse Buttons

Some platforms/terminals do not report mouse button for the MouseEventKind::Up and MouseEventKind::Drag events. MouseButton::Left is returned if we don't know which button was used.

Key Modifiers

Some platforms/terminals does not report all key modifiers combinations for all mouse event types. For example - macOS reports Ctrl + left mouse button click as a right mouse button click.

Crossterm.LibCrossterm.crossterm_MouseEventKind_Tag — Type

crossterm_MouseEventKind_Tag

A mouse event kind.

Platform-specific Notes

Mouse Buttons

Some platforms/terminals do not report mouse button for the MouseEventKind::Up and MouseEventKind::Drag events. MouseButton::Left is returned if we don't know which button was used.

Crossterm.LibCrossterm.crossterm_TerminalSize — Type

crossterm_TerminalSize

TerminalSize

Crossterm.LibCrossterm.crossterm_clear_last_error — Method

crossterm_clear_last_error()

Prototype

void crossterm_clear_last_error(void);

Crossterm.LibCrossterm.crossterm_cursor_disable_blinking — Method

crossterm_cursor_disable_blinking()

Disables blinking of the terminal cursor.

Prototype

int crossterm_cursor_disable_blinking(void);

Crossterm.LibCrossterm.crossterm_cursor_enable_blinking — Method

crossterm_cursor_enable_blinking()

Enables blinking of the terminal cursor.

Prototype

int crossterm_cursor_enable_blinking(void);

Crossterm.LibCrossterm.crossterm_cursor_hide — Method

crossterm_cursor_hide()

Hides the terminal cursor.

Prototype

int crossterm_cursor_hide(void);

Crossterm.LibCrossterm.crossterm_cursor_move_down — Method

crossterm_cursor_move_down(rows)

Moves the terminal cursor a given number of rows down.

Prototype

int crossterm_cursor_move_down(uint16_t rows);

Crossterm.LibCrossterm.crossterm_cursor_move_left — Method

crossterm_cursor_move_left(columns)

Moves the terminal cursor a given number of columns to the left.

Prototype

int crossterm_cursor_move_left(uint16_t columns);

Crossterm.LibCrossterm.crossterm_cursor_move_right — Method

crossterm_cursor_move_right(columns)

Moves the terminal cursor a given number of columns to the right.

Prototype

int crossterm_cursor_move_right(uint16_t columns);

Crossterm.LibCrossterm.crossterm_cursor_move_to_column — Method

crossterm_cursor_move_to_column(column)

Moves the terminal cursor to the given column on the current row.

Prototype

int crossterm_cursor_move_to_column(uint16_t column);

Crossterm.LibCrossterm.crossterm_cursor_move_to_next_line — Method

crossterm_cursor_move_to_next_line(n)

Moves the terminal cursor down the given number of lines and moves it to the first column.

Notes * This command is 1 based, meaning [crosstermcursormovetonext_line](@ref)(1) moves to the next line. * Most terminals default 0 argument to 1.

Prototype

int crossterm_cursor_move_to_next_line(uint16_t n);

Crossterm.LibCrossterm.crossterm_cursor_move_to_previous_line — Method

crossterm_cursor_move_to_previous_line(n)

Moves the terminal cursor up the given number of lines and moves it to the first column.

Prototype

int crossterm_cursor_move_to_previous_line(uint16_t n);

Crossterm.LibCrossterm.crossterm_cursor_move_to_row — Method

crossterm_cursor_move_to_row(row)

Moves the terminal cursor to the given row on the current column.

Prototype

int crossterm_cursor_move_to_row(uint16_t row);

Crossterm.LibCrossterm.crossterm_cursor_move_up — Method

crossterm_cursor_move_up(rows)

Moves the terminal cursor a given number of rows up.

Prototype

int crossterm_cursor_move_up(uint16_t rows);

Crossterm.LibCrossterm.crossterm_cursor_moveto — Method

crossterm_cursor_moveto(x, y)

Moves the terminal cursor to the given position (column, row).

Notes * Top left cell is represented as 0,0.

Prototype

int crossterm_cursor_moveto(uint16_t x, uint16_t y);

Crossterm.LibCrossterm.crossterm_cursor_position_get — Method

crossterm_cursor_position_get(pos)

Get cursor position (column, row)

Prototype

int crossterm_cursor_position_get(crossterm_CursorPosition *pos);

Crossterm.LibCrossterm.crossterm_cursor_position_set — Method

crossterm_cursor_position_set(pos)

Set cursor position (column, row)

Prototype

int crossterm_cursor_position_set(crossterm_CursorPosition pos);

Crossterm.LibCrossterm.crossterm_cursor_restore_position — Method

crossterm_cursor_restore_position()

Restores the saved terminal cursor position.

Prototype

int crossterm_cursor_restore_position(void);

Crossterm.LibCrossterm.crossterm_cursor_save_position — Method

crossterm_cursor_save_position()

Saves the current terminal cursor position.

Prototype

int crossterm_cursor_save_position(void);

Crossterm.LibCrossterm.crossterm_cursor_set_style — Method

crossterm_cursor_set_style(cursor_style)

Sets the style of the cursor.

Prototype

int crossterm_cursor_set_style(crossterm_CursorStyle cursor_style);

Crossterm.LibCrossterm.crossterm_cursor_set_style_blinking_bar — Method

crossterm_cursor_set_style_blinking_bar()

Sets the style of the cursor to a blinking bar.

Prototype

int crossterm_cursor_set_style_blinking_bar(void);

Crossterm.LibCrossterm.crossterm_cursor_set_style_blinking_block — Method

crossterm_cursor_set_style_blinking_block()

Sets the style of the cursor to a blinking block.

Prototype

int crossterm_cursor_set_style_blinking_block(void);

Crossterm.LibCrossterm.crossterm_cursor_set_style_blinking_underscore — Method

crossterm_cursor_set_style_blinking_underscore()

Sets the style of the cursor to a blinking underscore.

Prototype

int crossterm_cursor_set_style_blinking_underscore(void);

Crossterm.LibCrossterm.crossterm_cursor_set_style_default_user_shape — Method

crossterm_cursor_set_style_default_user_shape()

Sets the style of the cursor to default user shape.

Prototype

int crossterm_cursor_set_style_default_user_shape(void);

Crossterm.LibCrossterm.crossterm_cursor_set_style_steady_bar — Method

crossterm_cursor_set_style_steady_bar()

Sets the style of the cursor to a steady bar.

Prototype

int crossterm_cursor_set_style_steady_bar(void);

Crossterm.LibCrossterm.crossterm_cursor_set_style_steady_block — Method

crossterm_cursor_set_style_steady_block()

Sets the style of the cursor to a steady block.

Prototype

int crossterm_cursor_set_style_steady_block(void);

Crossterm.LibCrossterm.crossterm_cursor_set_style_steady_underscore — Method

crossterm_cursor_set_style_steady_underscore()

Sets the style of the cursor to a steady underscore.

Prototype

int crossterm_cursor_set_style_steady_underscore(void);

Crossterm.LibCrossterm.crossterm_cursor_show — Method

crossterm_cursor_show()

Shows the terminal cursor.

Prototype

int crossterm_cursor_show(void);

Crossterm.LibCrossterm.crossterm_disable_line_wrap — Method

crossterm_disable_line_wrap()

Disables line wrapping.

Prototype

int crossterm_disable_line_wrap(void);

Crossterm.LibCrossterm.crossterm_enable_line_wrap — Method

crossterm_enable_line_wrap()

Enables line wrapping.

Prototype

int crossterm_enable_line_wrap(void);

Crossterm.LibCrossterm.crossterm_enter_alternate_screen — Method

crossterm_enter_alternate_screen()

Enters alternate screen.

Prototype

int crossterm_enter_alternate_screen(void);

Crossterm.LibCrossterm.crossterm_event_disable_bracketed_paste — Method

crossterm_event_disable_bracketed_paste()

Disables bracketed paste mode.

Prototype

int crossterm_event_disable_bracketed_paste(void);

Crossterm.LibCrossterm.crossterm_event_disable_focus_change — Method

crossterm_event_disable_focus_change()

Disable focus event emission.

Prototype

int crossterm_event_disable_focus_change(void);

Crossterm.LibCrossterm.crossterm_event_disable_mouse_capture — Method

crossterm_event_disable_mouse_capture()

Disable mouse event capturing.

Prototype

int crossterm_event_disable_mouse_capture(void);

Crossterm.LibCrossterm.crossterm_event_enable_bracketed_paste — Method

crossterm_event_enable_bracketed_paste()

Enables bracketed paste mode.

It should be paired with [crossterm_event_disable_bracketed_paste] at the end of execution.

This is not supported in older Windows terminals without virtual terminal sequences.

Prototype

int crossterm_event_enable_bracketed_paste(void);

Crossterm.LibCrossterm.crossterm_event_enable_focus_change — Method

crossterm_event_enable_focus_change()

Enable focus event emission.

It should be paired with [crossterm_event_disable_focus_change] at the end of execution.

Focus events can be captured with [crossterm_event_read].

Prototype

int crossterm_event_enable_focus_change(void);

Crossterm.LibCrossterm.crossterm_event_enable_mouse_capture — Method

crossterm_event_enable_mouse_capture()

Enable mouse event capturing.

Prototype

int crossterm_event_enable_mouse_capture(void);

Crossterm.LibCrossterm.crossterm_event_poll — Method

crossterm_event_poll(secs, nanos)

Checks if there is an [Event] available.

Returns true if an [Event] is available otherwise it returns false.

true guarantees that subsequent call to the [crossterm_event_read] function won't block.

Arguments

timeout_secs - maximum waiting time for event availability * timeout_nanos - maximum waiting time for event availability

Prototype

int crossterm_event_poll(uint64_t secs, uint32_t nanos);

Crossterm.LibCrossterm.crossterm_event_pop_keyboard_enhancement_flags — Method

crossterm_event_pop_keyboard_enhancement_flags()

Disables extra kinds of keyboard events.

Prototype

int crossterm_event_pop_keyboard_enhancement_flags(void);

Crossterm.LibCrossterm.crossterm_event_push_keyboard_enhancement_flags — Method

crossterm_event_push_keyboard_enhancement_flags(flags)

Enables the kitty keyboard protocol, which adds extra information to keyboard events and removes ambiguity for modifier keys. It should be paired with [crossterm_pop_keyboard_enhancement_flags] at the end of execution.

Prototype

int crossterm_event_push_keyboard_enhancement_flags(uint8_t flags);

Crossterm.LibCrossterm.crossterm_event_read — Method

crossterm_event_read()

Reads a single [Event] as a UTF-8 json string.

This function blocks until an [Event] is available. Combine it with the [crossterm_event_poll] function to get non-blocking reads. User is responsible to free string. Use [crossterm_free_c_char] to free data

Prototype

const char *crossterm_event_read(void);

Crossterm.LibCrossterm.crossterm_free_c_char — Method

crossterm_free_c_char(s)

Frees data behind pointer to UTF-8 string allocated by this crate

Null character is stored in the last location of buffer.

Prototype

int crossterm_free_c_char(char *s);

Crossterm.LibCrossterm.crossterm_has_error — Method

crossterm_has_error()

Check whether error has been set.

Prototype

bool crossterm_has_error(void);

Crossterm.LibCrossterm.crossterm_last_error_length — Method

crossterm_last_error_length()

Calculate the number of bytes in the last error's error message including a trailing null character. If there are no recent error, then this returns 0.

Prototype

int crossterm_last_error_length(void);

Crossterm.LibCrossterm.crossterm_last_error_message — Method

crossterm_last_error_message()

Return most recent error message into a UTF-8 string buffer.

Null character is stored in the last location of buffer. Caller is responsible to memory associated with string buffer. Use [crossterm_free_c_char] to free data.

Prototype

const char *crossterm_last_error_message(void);

Crossterm.LibCrossterm.crossterm_leave_alternate_screen — Method

crossterm_leave_alternate_screen()

Leaves alternate screen.

Prototype

int crossterm_leave_alternate_screen(void);

Crossterm.LibCrossterm.crossterm_sleep — Method

crossterm_sleep(seconds)

Sleeps for seconds seconds

Prototype

void crossterm_sleep(double seconds);

Crossterm.LibCrossterm.crossterm_style_reset_color — Method

crossterm_style_reset_color()

Resets the colors back to default.

Prototype

int crossterm_style_reset_color(void);

Crossterm.LibCrossterm.crossterm_style_set_attribute — Method

crossterm_style_set_attribute(attr)

Sets an attribute.

See [Attribute] for more info.

Prototype

int crossterm_style_set_attribute(crossterm_Attribute attr);

Crossterm.LibCrossterm.crossterm_style_set_background_color — Method

crossterm_style_set_background_color(color)

Sets the the background color.

See [Color] for more info.

Prototype

int crossterm_style_set_background_color(crossterm_Color color);

Crossterm.LibCrossterm.crossterm_style_set_foreground_color — Method

crossterm_style_set_foreground_color(color)

Sets the the foreground color.

See [Color] for more info.

Prototype

int crossterm_style_set_foreground_color(crossterm_Color color);

Crossterm.LibCrossterm.crossterm_style_set_underline_color — Method

crossterm_style_set_underline_color(color)

Sets the the underline color.

See [Color] for more info.

Prototype

int crossterm_style_set_underline_color(crossterm_Color color);

Crossterm.LibCrossterm.crossterm_terminal_begin_synchronized_update — Method

crossterm_terminal_begin_synchronized_update()

Instructs the terminal emulator to begin a synchronized frame.

Notes

Commands must be executed/queued for execution otherwise they do nothing. * Use [crossterm_terminal_end_synchronized_update] command to leave the entered alternate screen.

When rendering the screen of the terminal, the Emulator usually iterates through each visible grid cell and renders its current state. With applications updating the screen a at higher frequency this can cause tearing.

This mode attempts to mitigate that.

When the synchronization mode is enabled following render calls will keep rendering the last rendered state. The terminal Emulator keeps processing incoming text and sequences. When the synchronized update mode is disabled again the renderer may fetch the latest screen buffer state again, effectively avoiding the tearing effect by unintentionally rendering in the middle a of an application screen update.

Prototype

int crossterm_terminal_begin_synchronized_update(void);

Crossterm.LibCrossterm.crossterm_terminal_clear — Method

crossterm_terminal_clear(ct)

Clear screen command.

Prototype

int crossterm_terminal_clear(crossterm_ClearType ct);

Crossterm.LibCrossterm.crossterm_terminal_disable_raw_mode — Method

crossterm_terminal_disable_raw_mode()

Disables raw mode.

Prototype

int crossterm_terminal_disable_raw_mode(void);

Crossterm.LibCrossterm.crossterm_terminal_enable_raw_mode — Method

crossterm_terminal_enable_raw_mode()

Enables raw mode.

Prototype

int crossterm_terminal_enable_raw_mode(void);

Crossterm.LibCrossterm.crossterm_terminal_end_synchronized_update — Method

crossterm_terminal_end_synchronized_update()

Instructs the terminal to end a synchronized frame.

Notes

Commands must be executed/queued for execution otherwise they do nothing. * Use [crossterm_terminal_begin_synchronized_update] to enter the alternate screen.

When rendering the screen of the terminal, the Emulator usually iterates through each visible grid cell and renders its current state. With applications updating the screen a at higher frequency this can cause tearing.

This mode attempts to mitigate that.

When the synchronization mode is enabled following render calls will keep rendering the last rendered state. The terminal Emulator keeps processing incoming text and sequences. When the synchronized update mode is disabled again the renderer may fetch the latest screen buffer state again, effectively avoiding the tearing effect by unintentionally rendering in the middle a of an application screen update.

Prototype

int crossterm_terminal_end_synchronized_update(void);

Crossterm.LibCrossterm.crossterm_terminal_scroll_down — Method

crossterm_terminal_scroll_down(n)

Scroll down command.

Prototype

int crossterm_terminal_scroll_down(unsigned short n);

Crossterm.LibCrossterm.crossterm_terminal_scroll_up — Method

crossterm_terminal_scroll_up(n)

Scroll up command.

Prototype

int crossterm_terminal_scroll_up(unsigned short n);

Crossterm.LibCrossterm.crossterm_terminal_set_size — Method

crossterm_terminal_set_size(columns, rows)

Sets the terminal buffer size (columns, rows).

Prototype

int crossterm_terminal_set_size(uint16_t columns, uint16_t rows);

Crossterm.LibCrossterm.crossterm_terminal_set_title — Method

crossterm_terminal_set_title(title)

Sets terminal title.

Safety

This function takes a raw pointer as argument. As such, the caller must ensure that: - The title pointer points to a valid null-terminated string. - This function borrows a slice to a valid null-terminated string and the memory referenced by title won't be deallocated or modified for the duration of the function call.. - The title pointer is correctly aligned and title points to an initialized memory.

If these conditions are not met, the behavior is undefined.

Prototype

int crossterm_terminal_set_title(const char *title);

Crossterm.LibCrossterm.crossterm_terminal_size — Method

crossterm_terminal_size(size)

Get terminal size

Prototype

int crossterm_terminal_size(crossterm_TerminalSize *size);